# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Ethnos.app Academic Bibliography API
  version: 2.0.0
  description: "

    REST API for academic bibliography: search, works, persons,
    organizations, venues, courses, citations, collaborations and health.


    Highlights

    - Public endpoints; selected admin/metrics routes require
    'x-access-key'.

    - Standard response envelope: { status, data, pagination?, meta? }.

    - Pagination supports page/limit and offset/limit.

    - Identifiers: DOI, ORCID, ROR, ISSN.


    Notes

    - Rate limiting is configurable via environment and enabled by default.

    - Performance and dataset size vary by deployment.

    \        "
  contact:
    name: Bruno Cesar Cunha Cruz, PhD Student
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  x-developer-orcid: 0000-0001-8652-2333
  x-institution: PPGAS/MN/UFRJ
servers:
  - url: https://api.ethnos.app
    description: Production API Server - Enterprise-ready academic research infrastructure
  - url: http://localhost:1210
    description: Development Server - Local testing and development environment
...
components:
  securitySchemes:
    XAccessKey:
      type: apiKey
      in: header
      name: x-access-key
      description: Internal access key required for protected endpoints (security,
        dashboard, health except /live)
  schemas:
...
    Error:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Resource not found
        code:
          type: string
          example: NOT_FOUND
        timestamp:
          type: string
          format: date-time
...
  responses:
...
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: error
            message: Invalid venue ID
            code: VALIDATION_ERROR
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: error
            message: Resource not found
            code: NOT_FOUND
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: error
            message: Internal server error
            code: INTERNAL_ERROR
...
paths:
...
  "/works/{id}/bibliography":
    get:
      summary: Get work bibliography usage
      description: Retrieve courses where this work is used in bibliography, with
        instructor information
      tags:
        - Works
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: Work ID
          example: 2684644
        - in: query
          name: reading_type
          schema:
            type: string
            enum:
              - REQUIRED
              - RECOMMENDED
              - SUPPLEMENTARY
              - OPTIONAL
          description: Filter by reading type
          example: RECOMMENDED
        - in: query
          name: year_from
          schema:
            type: integer
            minimum: 1900
            maximum: 2030
          description: Filter courses from this year
          example: 2020
        - in: query
          name: year_to
          schema:
            type: integer
            minimum: 1900
            maximum: 2030
          description: Filter courses up to this year
          example: 2025
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        "200":
          description: Work bibliography usage retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        course_id:
                          type: integer
                          example: 465
                        course_name:
                          type: string
                          example: Antropologia do Parentesco
                        course_year:
                          type: integer
                          example: 2025
                        program_id:
                          type: integer
                          example: 2
                        reading_type:
                          type: string
                          enum:
                            - REQUIRED
                            - RECOMMENDED
                            - SUPPLEMENTARY
                            - OPTIONAL
                          example: RECOMMENDED
                        instructor_count:
                          type: integer
                          example: 2
                        instructors:
                          type: string
                          example: João Silva; Maria Santos
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"